Introduce vnode IO_FILLORKILL flag and expand ioflag field
Adds the I/O flag IO_FILLORKILL that allows writes to reserve their
resources upfront thereby skipping several other points of blocking in
the stack and allowing the I/O sheduler to reject work under load.

To complete this I expanded the field to 64-bits to allow for more
flags.

Sponsored by: Netflix

+++ sys/fs/msdosfs/msdosfs_vnops.c
@@ -622,7 +622,7 @@ msdosfs_write(struct vop_write_args *ap)
 	int seqcount;
 	daddr_t bn, lastcn;
 	struct buf *bp;
-	int ioflag = ap->a_ioflag;
+	uint64_t ioflag = ap->a_ioflag;
 	struct uio *uio = ap->a_uio;
 	struct vnode *vp = ap->a_vp;
 	struct vnode *thisvp;
